pyproject.toml
[project]
name = "wp-python"
version = "0.1.0"
description = "A python version of wordpress rest api"
authors = [
    {name = "haoyue",email = "haoyue@coralera.org"}
]
readme = "README.md"
requires-python = ">=3.11,<4"
dependencies = [
    "httpx (>=0.28.1,<0.29.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "requests (>=2.32.4,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "rich (>=14.1.0,<15.0.0)",
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn[standard] (>=0.35.0,<0.36.0)",
    "gunicorn (>=23.0.0,<24.0.0)"
]

[tool.poetry]
packages = [{include = "wp_python", from = "src"}]


[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
